requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
libml @ git+https://github.com/remla25-team12/lib-ml.git@v0.1.5#egg=libml
joblib~=1.4.2
requests~=2.32.3
PyYAML==6.0.1
pylint==3.3.6
dvc~=3.59.2
dvc[gdrive]~=3.59.2
pytest>=7.0.0
pytest-cov>=4.0.0
psutil>=5.8.0
black==24.3.0
isort==5.13.2
autopep8==2.0.4
flake8==7.0.0